requirements.txt
torch
transformers==4.48.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.2.2
openai==1.12.0
python-dotenv==1.0.0
requests==2.32.3
beautifulsoup4==4.12.3
tmdbsimple==2.9.1
psutil==5.9.8
time-machine==2.14.1
memory-profiler==0.61.0
line-profiler==4.1.2
networkx
matplotlib
